+++ xenon-common/src/main/java/com/vmware/xenon/services/common/ConsistentHashingNodeSelectorService.java
@@ -489,7 +489,7 @@ public class ConsistentHashingNodeSelectorService extends StatelessService imple
         if (this.cachedGroupState == null) {
             op.fail(null);
         }
-        this.replicationUtility.replicateUpdate(this.cachedGroupState, op, body, response);
+        this.replicationUtility.replicateUpdate(this.cachedGroupState, op, body, response, this.cachedState.replicationQuorum);
     }
 
     /**
@@ -739,6 +739,21 @@ public class ConsistentHashingNodeSelectorService extends StatelessService imple
                 return;
             }
         }
+    }
+
+    @Override
+    public void updateReplicationQuorum(Operation op, int replicationQuorum) {
+        int replicationFactor = this.cachedState.replicationFactor != null ?
+                this.cachedState.replicationFactor.intValue() : this.cachedGroupState.nodes.size();
+        if (replicationQuorum > replicationFactor) {
+            String errorMsg = String.format(
+                    "replicationQuorum %d > replicationFactor %d", replicationQuorum, replicationFactor);
+            op.fail(new IllegalArgumentException(errorMsg));
+            return;
+        }
+        logInfo("replicationQuorum update from %d to %d", this.cachedState.replicationQuorum, replicationQuorum);
+        this.cachedState.replicationQuorum = replicationQuorum;
+        op.complete();
     }
 
     @Override